Add files via upload
AFT_modular_capacities.py -  is the implementation file for AFT with modular capacities problem.
germanNetwork_inputdata.xlsx - is the input data file for the German Network with the traffix matrix 100%.
Results_AffineFlowThinning_modular_capacities100%_tolerance(10e-1).xlsx - in this file you can find the results for AFT problem with Modular Capacities when the traffic matrix is 100%  and the tolerance is set to 10e-1. It is also calculated the working capacity and spare capacity.
Results_AffineFlowThinning_modular_capacities100%_tolerance(10e-3).xlsx - in this file you can find the results for AFT problem with Modular Capacities when the traffic matrix is 100%  and the tolerance is set to 10e-3. It is also calculated the working capacity and spare capacity.
Results_AffineFlowThinning_modular_capacities100%_tolerance(10e-6).xlsx - in this file you can find the results for AFT problem with Modular Capacities when the traffic matrix is 100%  and the tolerance is set to 10e-6. It is also calculated the working capacity and spare capacity.
Results_AffineFlowThinning_modular_capacities100%_tolerance(10e-9).xlsx - in this file you can find the results for AFT problem with Modular Capacities when the traffic matrix is 100%  and the tolerance is set to 10e-9. It is also calculated the working capacity and spare capacity.
Evaluation_tolerances.xlsx - contains calculations for  working and spare capacity for AFT with modular capacities when the traffic matrix is 100% for different tolerances.